+++ common/src/main/java/com/frostwire/search/torrent/TorrentCrawledAlbumSearchResult.java
@@ -28,12 +28,11 @@ import java.util.List;
  * @author aldenml
  */
 public final class TorrentCrawledAlbumSearchResult extends AbstractCrawledSearchResult<TorrentCrawlableSearchResult> implements TorrentSearchResult {
-
     private final String artist;
     private final String album;
     private final List<TorrentItemSearchResult> items;
     private final String displayName;
-    private final long size;
+    private final double size;
 
     public TorrentCrawledAlbumSearchResult(TorrentCrawlableSearchResult sr, String artist, String album, List<TorrentItemSearchResult> items) {
         super(sr);
@@ -67,7 +66,7 @@ public final class TorrentCrawledAlbumSearchResult extends AbstractCrawledSearch
     }
 
     @Override
-    public long getSize() {
+    public double getSize() {
         return size;
     }
 
@@ -111,17 +110,14 @@ public final class TorrentCrawledAlbumSearchResult extends AbstractCrawledSearch
         if (StringUtils.isNotBlank(artist)) {
             result = artist + " - " + album;
         }
-
         return result;
     }
 
     private long buildSize(List<TorrentItemSearchResult> items) {
         long r = 0;
-
         for (TorrentItemSearchResult sr : items) {
             r += sr.getSize();
         }
-
         return r;
     }
 }
