+++ common/src/main/java/com/frostwire/search/torrent/TorrentCrawledSearchResult.java
@@ -27,13 +27,12 @@ import org.apache.commons.io.FilenameUtils;
  * @author aldenml
  */
 public final class TorrentCrawledSearchResult extends AbstractCrawledSearchResult<TorrentCrawlableSearchResult> implements TorrentItemSearchResult {
-
     private final TorrentInfo ti;
     private final int fileIndex;
     private final String filePath;
     private final String displayName;
     private final String filename;
-    private final long size;
+    private final double size;
 
     public TorrentCrawledSearchResult(TorrentCrawlableSearchResult sr, TorrentInfo ti, int fileIndex, String filePath, long fileSize) {
         super(sr);
@@ -69,7 +68,7 @@ public final class TorrentCrawledSearchResult extends AbstractCrawledSearchResul
     }
 
     @Override
-    public long getSize() {
+    public double getSize() {
         return size;
     }
